honey I shrunk the kids
The game scene is overhauled

Kareena's assets are all pulled in, and I shrunk down every object in the game scene to translate to realistic meters. (1 unit in unity translates to 1 real world meter, and can easily be checked by adding an empty cube to the scene)

Unfortunately I accidentally deleted the playermodels in the process, but we should be getting all new ones shortly. The player objects work, but are invisible ghosts.

Oh yeah and if you wonder what that blank room is for, it's just a set-piece to make the doorway not look like it leads to oblivion.